+++ bot/utilities/cartography/constants.py
@@ -1,0 +1,15 @@
+SKELD_FILENAME = "SKELD.json"
+MIRAHQ_FILENAME = "MIRAHQ.json"
+POLUS_FILENAME = "POLUS.json"
+RED_PLAYER_MARKER_FILENAME = "RED_PLAYER_MARKER.jpg"
+BROWN_PLAYER_MARKER_FILENAME = "BROWN_PLAYER_MARKER.jpg"
+ORANGE_PLAYER_MARKER_FILENAME = "ORANGE_PLAYER_MARKER.jpg"
+YELLOW_PLAYER_MARKER_FILENAME = "YELLOW_PLAYER_MARKER.jpg"
+LIME_PLAYER_MARKER_FILENAME = "LIME_PLAYER_MARKER.jpg"
+GREEN_PLAYER_MARKER_FILENAME = "GREEN_PLAYER_MARKER.jpg"
+CYAN_PLAYER_MARKER_FILENAME = "CYAN_PLAYER_MARKER.jpg"
+BLUE_PLAYER_MARKER_FILENAME = "BLUE_PLAYER_MARKER.jpg"
+PURPLE_PLAYER_MARKER_FILENAME = "PURPLE_PLAYER_MARKER.jpg"
+PINK_PLAYER_MARKER_FILENAME = "PINK_PLAYER_MARKER.jpg"
+WHITE_PLAYER_MARKER_FILENAME = "WHITE_PLAYER_MARKER.jpg"
+BLACK_PLAYER_MARKER_FILENAME = "BLACK_PLAYER_MARKER.jpg"
