+++ bot/utilities/screen/constants.py
@@ -1,0 +1,13 @@
+# frames per second
+ONE_SECOND = 1
+FPS_1 = ONE_SECOND
+FPS_2 = ONE_SECOND / 2
+FPS_4 = ONE_SECOND / 4
+FPS_8 = ONE_SECOND / 8
+FPS_16 = ONE_SECOND / 16
+FPS_24 = ONE_SECOND / 24
+FPS_30 = ONE_SECOND / 30
+FPS_32 = ONE_SECOND / 32
+FPS_60 = ONE_SECOND / 60
+FPS_120 = ONE_SECOND / 120
+FPS_UNLIMITED = 0
